refactor: Fix type definition
