+++ src/app/blog/page.tsx
@@ -3,7 +3,7 @@ import { Metadata } from 'next';
 import BlogField from '@/components/model/blogField';
 import { BLOG } from '@/const/seo';
 import { getPosts } from '@/features/common/getPosts';
-import { Posts } from '@/types/post';
+import { Post } from '@/types/post';
 
 const siteName = 'zoniha';
 const description = BLOG.DESCRIPTION;
@@ -23,7 +23,7 @@ export const metadata: Metadata = {
 };
 
 const Home = async () => {
-  const posts: Posts = await getPosts();
+  const posts: Post[] = await getPosts();
 
   return (
     <>
