+++ src/types/formInputs.ts
@@ -1,7 +1,7 @@
-export type FormInputs = {
+export interface FormInputs {
   name: string;
   email: string;
   subject: string;
   message: string;
   [key: string]: string;
-};
+}
