+++ src/types/post.ts
@@ -1,21 +1,19 @@
-export type Post = {
+export interface Post {
   key: string;
   link: string;
   title: string;
   createdAt: string;
   tags: string[];
-};
+}
 
-export type Posts = Post[];
-
-export type ZennPost = {
+export interface ZennPost {
   [key: string]: any;
-};
+}
 
-export type BlogPost = {
+export interface BlogPost {
   __typename: string;
   createdAt: string;
   title: string;
   slug: string;
   id: string;
-};
+}
